Configurable ports for query & index service
Read the ports these bind to in the cluster from the values.yaml, rather
than using the ports that the underlying Content instances uses (though
only for ACI; index/service ports need to match what is advertised in
the GetStatus response, i.e. what is configured in the server itself).

Default to 9060 (query service) and 9070 (index service), to match the
defaults for DAH/DIH in the distributed-idol chart.