US#456048 Make IDOL Admin work through Ingress (#48)
Updated idol-library to expose aci, service and index ports so that they
can be accessed via idol admin, e.g. for single-content we can now
view the config file and index documents via idol admin. This only
applies to nginx as it doesn't seem possible via haproxy openshift. 

Also updated ingress templating to make it easier to implement new
paths. Added "portmapping" and "paths" variables to _ingress.tpl. The
former is a dictionary of <port name>:<port path> entries, which are
parsed by the template, and the latter is a list of full path specifications
which must be set in the relevant chart's ingress.yaml file (this last bit
is necessary since helm doesn't merge arrays when templating, so the
full paths array must be specified in idol-library.ingress.base).

Updated most charts to use the above functionality. Idol-nifi and
idol-find have not been updated for now as these use
idol-library.labels instead of idol-library.ingress, so updating these
will be a bit more involved. Idol-nifi version has still been updated as
it is using latest versions of idol-library and idol-licenseserver. 

Also removed idol-licenseserver's dependency on idol-library as
usage was minimal.

---------

Co-authored-by: daniel-lau-mf <36928012+daniel-lau-mf@users.noreply.github.com>
Reviewed-by: daniel-lau-mf <36928012+daniel-lau-mf@users.noreply.github.com>
Signed-off-by: hasan-mahmood-opentext <156795977+hasan-mahmood-opentext@users.noreply.github.com>