Generate sensitive props key & trust/keystore pass (#59)
If not specified in the values.yaml, generate default values for:
- nifi.security.keystorePasswd
- nifi.security.truststorePasswd
- nifi.sensitive.props.key

Move nifi generated keys to their own configmap. This prevents the case where if the user doesn't specify an explicit key, then any values change that affects the configmap resulting in the sensitive props key being rewritten and breaking the cluster. By moving the keys to their own configmap, this should limit the scope of any object update