US#493002 Ingress tls support (#60)
Added tls support for ingress. This is specified by a secretName
parameter, which will enable tls if non-empty. There are also crt and key
parameters if the tls secret is not generated beforehand. If secretName
is specified, but crt and key are left empty,  then we assume that the
secret already exists and attempt to use it in the ingress resource. If they
are specified, then we create a new secret using the provided (base64-
encoded) certificate and key.

Most of the changes are in idol-library, for which the updated release
has been added to the release directory. Other charts versions have
been updated as appropriate, but releases have not yet been added.

Idol-find is also now up to date with the latest idol-library.

Implements: https://internal.almoctane.com/ui/entity-navigation?p=161001/2002&entityType=work_item&id=493002
Co-authored-by: daniel-lau-mf <36928012+daniel-lau-mf@users.noreply.github.com>
Reviewed-by: daniel-lau-mf <36928012+daniel-lau-mf@users.noreply.github.com>
Signed-off-by: hasan-mahmood-opentext <156795977+hasan-mahmood-opentext@users.noreply.github.com>