idol-nifi: call pg-enable-services on child groups (#101)
The NiFi Toolkit pg-enable-services does not check that services started correctly if they are defined in a child process group. If the service defined in the child process group is able to be started, all is well, but if the service is still validating, or fails to start for any reason, the service is left disabled.
To work around this, after calling pg-enable-services on the  root process group for imported flows, recursively call pg-enable-services on all descendent process groups as well.