Fix gateway ingress for multi-component deployments
We can only really have one Gateway for a deployment, otherwise
host/port clashes occur and only one Gateway will successfully
bind. Remove Gateway from the ingress templating, thereby treating
it in the same manner as an nginx controller (i.e. we assume the
cluster administrator has already set one up, and we don't need to
create one ourselves).

Allow gateway name and namespace to be specified in the values.yaml
files. Namespace is optional and is only required if the Gateway exists
in a different namespace to the deployment.

Update several charts, having tested that Gateway ingresses work for
them with the changes.

Signed-off-by: lewis-mackay-mf