Reinstate service/port values for subcomponents
It makes sense to keep them here, since we won't have access to
subchart values if a subchart is disabled (e.g. in idol-stack).

Fix issue with NiFi gateway ingresses in passing; the fix when
adding app/JDK labels for normal ingresses needs to be made for
gateway ingresses too.

Update idol-stack to use the new Find.

Signed-off-by: lewis-mackay-mf